Refactor manager to separate core functions
