+++ manager-rs/dauth-service/src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth_vectors;
+pub mod confirm_keys;
+pub mod users;
